Refactor asset preview metadata derivation
